+++ ECommerceMVC/ECommerceMVC/Data/VChiTietHoaDon.cs
@@ -11,11 +11,11 @@ public partial class VChiTietHoaDon
 
     public int MaHh { get; set; }
 
-    public double DonGia { get; set; }
+    public decimal DonGia { get; set; }
 
     public int SoLuong { get; set; }
 
-    public double GiamGia { get; set; }
+    public decimal GiamGia { get; set; }
 
     public string TenHh { get; set; } = null!;
 }
