+++ ECommerceMVC/ECommerceMVC/ViewModels/CartItem.cs
@@ -5,8 +5,8 @@
 		public int MaHh { get; set; }
 		public string Hinh { get; set; }
 		public string TenHH { get; set; }
-		public double DonGia { get; set; }
+		public decimal DonGia { get; set; }
 		public int SoLuong { get; set; }
-		public double ThanhTien => SoLuong * DonGia;
+		public decimal ThanhTien => SoLuong * DonGia;
 	}
 }
